Move each snippet to its own file, with includes and client init This makes each snippet fully copy-pasteable Some additional cleanup
