+++ iam/api/Access/AddBinding.cs
@@ -1,4 +1,4 @@
-// Copyright 2018 Google Inc.
+﻿// Copyright 2018 Google Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,20 +12,21 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-using System;
-using Xunit;
+// [START iam_modify_policy_add_role]
+using System.Collections.Generic;
+using Google.Apis.CloudResourceManager.v1.Data;
 
-namespace GoogleCloudSamples
+public partial class AccessManager
 {
-    public class GrantableRolesTest
+    public static Policy AddBinding(Policy policy, string role, string member)
     {
-        [Fact]
-        public void TestGrantableRoles()
+        var binding = new Binding
         {
-            string project = Environment.GetEnvironmentVariable("GOOGLE_PROJECT_ID");
-            string resource = "//cloudresourcemanager.googleapis.com/projects/" + project;
-
-            GrantableRoles.ViewGrantableRoles(resource);
-        }
+            Role = role,
+            Members = new List<string> { member }
+        };
+        policy.Bindings.Add(binding);
+        return policy;
     }
 }
+// [END iam_modify_policy_add_role]
