+++ iam/api/CustomRoles/RolesCreate.cs
@@ -1,0 +1,52 @@
+// Copyright 2018 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// [START iam_create_role]
+using System;
+using System.Collections.Generic;
+using Google.Apis.Auth.OAuth2;
+using Google.Apis.Iam.v1;
+using Google.Apis.Iam.v1.Data;
+
+public partial class CustomRoles
+{
+    public static Role CreateRole(string name, string projectId, string title,
+        string description, IList<string> permissions, string stage)
+    {
+        var credential = GoogleCredential.GetApplicationDefault()
+            .CreateScoped(IamService.Scope.CloudPlatform);
+        var service = new IamService(new IamService.Initializer
+        {
+            HttpClientInitializer = credential
+        });
+
+        var role = new Role
+        {
+            Title = title,
+            Description = description,
+            IncludedPermissions = permissions,
+            Stage = stage
+        };
+        var request = new CreateRoleRequest
+        {
+            Role = role,
+            RoleId = name
+        };
+        role = service.Projects.Roles.Create(request,
+            "projects/" + projectId).Execute();
+        Console.WriteLine("Created role: " + role.Name);
+        return role;
+    }
+}
+// [END iam_create_role]
