--- iam/api/GrantableRoles/GrantableRoles.csproj
@@ -1,12 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.0</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Google.Apis.Iam.v1" Version="1.36.1.1429" />
-  </ItemGroup>
-
-</Project>
