+++ iam/api/ServiceAccounts/Main.cs
@@ -1,0 +1,49 @@
+// Copyright 2018 Google Inc.
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using CommandLine;
+
+public partial class ServiceAccounts
+{
+    public static void Main(string[] args)
+    {
+        Parser.Default.ParseArguments<
+            CreateServiceAccountOptions,
+            ListServiceAccountOptions,
+            RenameServiceAccountOptions,
+            DeleteServiceAccountOptions
+            >(args).MapResult(
+            (CreateServiceAccountOptions x) =>
+            {
+                CreateServiceAccount(x.ProjectId, x.Name, x.DisplayName);
+                return 0;
+            },
+            (ListServiceAccountOptions x) =>
+            {
+                ListServiceAccounts(x.ProjectId);
+                return 0;
+            },
+            (RenameServiceAccountOptions x) =>
+            {
+                RenameServiceAccount(x.Email, x.DisplayName);
+                return 0;
+            },
+            (DeleteServiceAccountOptions x) =>
+            {
+                DeleteServiceAccount(x.Email);
+                return 0;
+            },
+            error => 1);
+    }
+}
