+++ iam/api/ServiceAccounts/Options.cs
@@ -14,42 +14,39 @@
 
 using CommandLine;
 
-namespace GoogleCloudSamples
+[Verb("create", HelpText = "Creates a service account.")]
+internal class CreateServiceAccountOptions
 {
-    [Verb("create", HelpText = "Creates a service account.")]
-    internal class CreateServiceAccountOptions
-    {
-        [Option("project-id", HelpText = "The project ID.", Required = true)]
-        public string ProjectId { get; set; }
+    [Option("project-id", HelpText = "The project ID.", Required = true)]
+    public string ProjectId { get; set; }
 
-        [Option("name", HelpText = "The service account's name.", Required = true)]
-        public string Name { get; set; }
+    [Option("name", HelpText = "The service account's name.", Required = true)]
+    public string Name { get; set; }
 
-        [Option("display-name", HelpText = "The service account's friendly display name.")]
-        public string DisplayName { get; set; }
-    }
+    [Option("display-name", HelpText = "The service account's friendly display name.")]
+    public string DisplayName { get; set; }
+}
 
-    [Verb("list", HelpText = "Lists service accounts.")]
-    internal class ListServiceAccountOptions
-    {
-        [Option("project-id", HelpText = "The project ID.", Required = true)]
-        public string ProjectId { get; set; }
-    }
+[Verb("list", HelpText = "Lists service accounts.")]
+internal class ListServiceAccountOptions
+{
+    [Option("project-id", HelpText = "The project ID.", Required = true)]
+    public string ProjectId { get; set; }
+}
 
-    [Verb("rename", HelpText = "Updates a service account's display name.")]
-    internal class RenameServiceAccountOptions
-    {
-        [Option("email", HelpText = "The service account email.", Required = true)]
-        public string Email { get; set; }
+[Verb("rename", HelpText = "Updates a service account's display name.")]
+internal class RenameServiceAccountOptions
+{
+    [Option("email", HelpText = "The service account email.", Required = true)]
+    public string Email { get; set; }
 
-        [Option("display-name", HelpText = "The service account's new friendly display name.", Required = true)]
-        public string DisplayName { get; set; }
-    }
+    [Option("display-name", HelpText = "The service account's new friendly display name.", Required = true)]
+    public string DisplayName { get; set; }
+}
 
-    [Verb("delete", HelpText = "Deletes a service account.")]
-    internal class DeleteServiceAccountOptions
-    {
-        [Option("email", HelpText = "The service account's email.", Required = true)]
-        public string Email { get; set; }
-    }
+[Verb("delete", HelpText = "Deletes a service account.")]
+internal class DeleteServiceAccountOptions
+{
+    [Option("email", HelpText = "The service account's email.", Required = true)]
+    public string Email { get; set; }
 }
